requirement.txt
flask==2.3.3
openai==1.64.0
python-dotenv==1.0.0
gunicorn==21.2.0